+++ aioutils/__init__.py
@@ -4,4 +4,4 @@ from .yielder import Yielder, OrderedYielder, yielding, ordered_yielding
 
 __all__ = ['Pool', 'Group', 'Bag', 'OrderedBag',
            'Yielder', 'OrderedYielder', 'yielding', 'ordered_yielding']
-__version__ = '0.3.7'
+__version__ = '0.3.8'